requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
chardet==5.1.0
click==8.1.6
colorama==0.4.6
decorator==5.1.1
Django==4.2.3
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap-v5==1.0.11
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-intl-tel-input==0.3.1
django-phonenumber-field==7.1.0
django-recaptcha==3.0.0
django-tinymce==3.6.1
django-utils-six==2.0
executing==1.2.0
fontawesomefree==6.4.0
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pathspec==0.11.1
phonenumbers==8.13.16
pickleshare==0.7.5
pip-check-reqs==2.4.4
prompt-toolkit==3.0.39
pure-eval==0.2.2
pydantic==2.0.2
pydantic_core==2.1.2
Pygments==2.15.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6
